chore: moved back to TypeScript because existing tests aren't currently set up for TypeScript
